refactor: fix typo in package name coudb -> couchdb
