+++ src/main/java/de/bsi/secvisogram/csaf_cms_backend/couchdb/CouchDbService.java
@@ -1,4 +1,4 @@
-package de.bsi.secvisogram.csaf_cms_backend.coudb;
+package de.bsi.secvisogram.csaf_cms_backend.couchdb;
 
 import static de.bsi.secvisogram.csaf_cms_backend.json.AdvisoryJsonService.*;
