+++ src/main/java/de/bsi/secvisogram/csaf_cms_backend/couchdb/DatabaseException.java
@@ -1,4 +1,4 @@
-package de.bsi.secvisogram.csaf_cms_backend.coudb;
+package de.bsi.secvisogram.csaf_cms_backend.couchdb;
 
 public class DatabaseException extends Exception {
