+++ src/main/java/de/bsi/secvisogram/csaf_cms_backend/rest/AdvisoryController.java
@@ -4,8 +4,8 @@ package de.bsi.secvisogram.csaf_cms_backend.rest;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import de.bsi.secvisogram.csaf_cms_backend.SecvisogramApplication;
-import de.bsi.secvisogram.csaf_cms_backend.coudb.CouchDbService;
-import de.bsi.secvisogram.csaf_cms_backend.coudb.DatabaseException;
+import de.bsi.secvisogram.csaf_cms_backend.couchdb.CouchDbService;
+import de.bsi.secvisogram.csaf_cms_backend.couchdb.DatabaseException;
 import de.bsi.secvisogram.csaf_cms_backend.json.AdvisoryJsonService;
 import de.bsi.secvisogram.csaf_cms_backend.model.DocumentTrackingStatus;
 import de.bsi.secvisogram.csaf_cms_backend.model.ExportFormat;
